+++ zerver/tests/test_upload_local.py
@@ -76,7 +76,8 @@ class LocalStorageTest(UploadSerializeMixin, ZulipTestCase):
         result = self.client_post("/json/user_uploads", {"file": fp})
 
         response_dict = self.assert_json_success(result)
-        path_id = re.sub(r"/user_uploads/", "", response_dict["uri"])
+        self.assertEqual(response_dict["uri"], response_dict["url"])
+        path_id = re.sub(r"/user_uploads/", "", response_dict["url"])
 
         assert settings.LOCAL_FILES_DIR is not None
         file_path = os.path.join(settings.LOCAL_FILES_DIR, path_id)